+++ Branch-SDK/src/androidTest/java/io/branch/referral/BranchTest.java
@@ -2,7 +2,8 @@ package io.branch.referral;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import androidx.test.InstrumentationRegistry;
+
+import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Assert;
@@ -22,7 +23,7 @@ public class BranchTest {
 
     @Before
     public void setUp() {
-        mContext = InstrumentationRegistry.getContext();
+        mContext = ApplicationProvider.getApplicationContext();
     }
 
     @After