+++ Branch-SDK/src/main/java/io/branch/referral/PrefHelper.java
@@ -943,21 +943,24 @@ public class PrefHelper {
     }
 
     /**
-     * <p>Sets the latd attribution window</p>
+     * <p>Sets the latd attribution window, if attributionWindow is null, the saved latd attribution
+     * window value will be deleted.</p>
      *
-     * @param attributonWindow A {@link Integer} value containing the current attribution window passed
+     * @param attributionWindow An {@link Integer} value containing the current attribution window passed
      */
-    public void setLATDAttributonWindow(int attributonWindow){
-        setInteger(KEY_LATD_ATTRIBUTION_WINDOW, attributonWindow);
+    public void setLATDAttributionWindow(int attributionWindow){
+        setInteger(KEY_LATD_ATTRIBUTION_WINDOW, attributionWindow);
     }
 
     /**
      * <p>Gets the latd attribution window</p>
      *
-     * @return attributonWindow A {@link Integer} value containing the current attribution window set
+     * @return attributionWindow An {@link Integer} value containing the current attribution window or null
      */
-    public int getLATDAttributonWindow(){
-        return getInteger(KEY_LATD_ATTRIBUTION_WINDOW);
+    @SuppressWarnings("WeakerAccess")
+    public int getLATDAttributionWindow(){
+        return getInteger(KEY_LATD_ATTRIBUTION_WINDOW,
+                ServerRequestGetLATD.defaultAttributionWindow);
     }
     
     // ALL GENERIC CALLS
@@ -976,6 +979,25 @@ public class PrefHelper {
         String[] stringArr = list.split(",");
         Collections.addAll(strings, stringArr);
         return strings;
+    }
+
+    /**
+     * <p>A basic method that returns a {@link Boolean} indicating whether some preference exists.</p>
+     *
+     * @param key A {@link String} value containing the key to reference.
+     * @return A {@link Boolean} indicating whether some preference exists.
+     */
+    public boolean hasPrefValue(String key) {
+        return prefHelper_.appSharedPrefs_.contains(key);
+    }
+
+    /**
+     * <p>A basic method to remove some preference value.</p>
+     *
+     * @param key A {@link String} value containing the key to the value that's to be deleted.
+     */
+    public void removePrefValue(String key) {
+        prefHelper_.prefsEditor_.remove(key).apply();
     }
     
     /**