+++ Branch-SDK/src/main/java/io/branch/referral/ServerRequest.java
@@ -182,6 +182,7 @@ public abstract class ServerRequest {
      */
     protected void setPost(JSONObject post) throws JSONException {
         params_ = post;
+
         if (getBranchRemoteAPIVersion() == BRANCH_API_VERSION.V1) {
             DeviceInfo.getInstance().updateRequestWithV1Params(this, params_);
         } else {
@@ -189,10 +190,6 @@ public abstract class ServerRequest {
                 JSONObject userDataObj = new JSONObject();
                 params_.put(Defines.Jsonkey.UserData.getKey(), userDataObj);
                 DeviceInfo.getInstance().updateRequestWithV2Params(this, context_, prefHelper_, userDataObj);
-
-                if (getBranchRemoteAPIVersion() == BRANCH_API_VERSION.V1_LATD) {
-                    DeviceInfo.getInstance().updateRequestWithAttributionWindow(prefHelper_, userDataObj);
-                }
             } catch (JSONException ignored) {}
         }
     }